+++ sdk/identity/identity/src/credentials/interactiveBrowserCredentialOptions.ts
@@ -35,6 +35,12 @@ export type InteractiveBrowserCredentialOptions = TokenCredentialOptions &
      * The client (application) ID of an App Registration in the tenant.
      */
     clientId?: string;
+
+    /**
+     * loginHint allows a user name to be pre-selected for interactive logins.
+     * Setting this option skips the account selection prompt and immediately attempts to login with the specified account.
+     */
+    loginHint?: string;
   };
 
 /**
@@ -66,4 +72,10 @@ export type InteractiveBrowserCredentialBrowserOptions = TokenCredentialOptions 
      *
      */
     loginStyle?: BrowserLoginStyle;
+
+    /**
+     * loginHint allows a user name to be pre-selected for interactive logins.
+     * Setting this option skips the account selection prompt and immediately attempts to login with the specified account.
+     */
+    loginHint?: string;
   };