+++ sdk/identity/identity/src/msal/browserFlows/msalAuthCode.ts
@@ -19,6 +19,7 @@ const redirectHash = self.location.hash;
  */
 export class MSALAuthCode extends MsalBrowser {
   protected app: msalBrowser.PublicClientApplication;
+  private loginHint?: string;
 
   /**
    * Sets up an MSAL object based on the given parameters.
@@ -28,6 +29,7 @@ export class MSALAuthCode extends MsalBrowser {
    */
   constructor(options: MsalBrowserFlowOptions) {
     super(options);
+    this.loginHint = options.loginHint;
 
     this.msalConfig.cache = {
       cacheLocation: "sessionStorage",
@@ -120,8 +122,9 @@ To work with multiple accounts for the same Client ID and Tenant ID, please prov
    */
   public async login(scopes: string | string[] = []): Promise<AuthenticationRecord | undefined> {
     const arrayScopes = Array.isArray(scopes) ? scopes : [scopes];
-    const loginRequest = {
-      scopes: arrayScopes
+    const loginRequest: msalBrowser.RedirectRequest = {
+      scopes: arrayScopes,
+      loginHint: this.loginHint
     };
     switch (this.loginStyle) {
       case "redirect": {
@@ -189,6 +192,7 @@ To work with multiple accounts for the same Client ID and Tenant ID, please prov
       authority: this.msalConfig.auth.authority!,
       correlationId: options?.correlationId,
       account: publicToMsal(account),
+      loginHint: this.loginHint,
       scopes
     };