+++ sdk/identity/identity/src/msal/nodeFlows/msalOpenBrowser.ts
@@ -18,6 +18,7 @@ import { CredentialUnavailableError } from "../../client/errors";
  */
 export interface MSALOpenBrowserOptions extends MsalNodeOptions {
   redirectUri: string;
+  loginHint?: string;
 }
 
 /**
@@ -37,11 +38,13 @@ export class MsalOpenBrowser extends MsalNode {
   private redirectUri: string;
   private port: number;
   private hostname: string;
+  private loginHint?: string;
 
   constructor(options: MSALOpenBrowserOptions) {
     super(options);
     this.logger = credentialLogger("NodeJS MSAL Open Browser");
     this.redirectUri = options.redirectUri;
+    this.loginHint = options.loginHint;
 
     const url = new URL(this.redirectUri);
     this.port = parseInt(url.port);
@@ -185,7 +188,8 @@ export class MsalOpenBrowser extends MsalNode {
   private async openAuthCodeUrl(scopeArray: string[]): Promise<void> {
     const authCodeUrlParameters: msalNode.AuthorizationUrlRequest = {
       scopes: scopeArray,
-      redirectUri: this.redirectUri
+      redirectUri: this.redirectUri,
+      loginHint: this.loginHint
     };
 
     const response = await this.publicApp!.getAuthCodeUrl(authCodeUrlParameters);